+++ guillotina/renderers.py
@@ -1,4 +1,3 @@
-from aiohttp.helpers import sentinel
 from aiohttp.web import Response as aioResponse
 from datetime import datetime
 from guillotina import configure
@@ -49,21 +48,6 @@ class GuillotinaJSONEncoder(json.JSONEncoder):
 
 # b/w compat import
 PServerJSONEncoder = GuillotinaJSONEncoder
-
-
-def json_response(data=sentinel, *, text=None, body=None, status=200,
-                  reason=None, headers=None, content_type='application/json',
-                  dumps=json.dumps):
-    if data is not sentinel:
-        if text or body:
-            raise ValueError(
-                "only one of data, text, or body should be specified"
-            )
-        else:
-            text = dumps(data, cls=GuillotinaJSONEncoder)
-    return aioResponse(
-        text=text, body=body, status=status, reason=reason,
-        headers=headers, content_type=content_type)
 
 
 @configure.adapter(for_=IRequest, provides=IRenderFormats)
@@ -127,19 +111,21 @@ class RendererJson(Renderer):
         else:
             # Not a Response object, don't convert
             return value
+        if isinstance(json_value, aioResponse):
+            # not actually json
+            return json_value
+
         # Framing of options
         frame = self.request.get('frame')
         frame = self.request.query['frame'] if 'frame' in self.request.query else ''
         if frame:
             framer = query_adapter(self.request, IFrameFormatsJson, frame)
             json_value = await apply_coroutine(framer, json_value)
-        resp = json_response(json_value)
+        resp = aioResponse(text=json.dumps(json_value, cls=GuillotinaJSONEncoder))
         resp.headers.update(headers)
         resp.headers.update(
             {'Content-Type': 'application/json'})
         resp.set_status(status)
-        # Actions / workflow / roles
-
         return resp